% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syn_import.R
\name{syn_import}
\alias{syn_import}
\title{Wrapping from the base function read.csv.}
\usage{
syn_import(file = "")
}
\arguments{
\item{file}{a file name as the export file. (notice: add th subtitle .csv)}
}
\value{
a dataframe
}
\description{
Wrapping from the base function read.csv.
}
\details{

}
\examples{
DTF <- syn_import(file = 'test1124.csv')


}
