% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syn_replace.R
\name{syn_replace}
\alias{syn_replace}
\title{Replace synonyms by importing a csv file.}
\usage{
syn_replace(df, data, tag)
}
\arguments{
\item{df}{the edited file you import which includes synonyms and replaced terms}

\item{data}{a dataset the original biblio data}

\item{tag}{a field-tag, i.g. 'ID' or 'DE'.}
}
\value{
a dataframe
}
\description{
Replace synonyms by importing a csv file.
}
\details{

}
\examples{
data(scientometrics, package = "bibliometrixData")
scientometrics$ID <- syn_replace(
  df = DTF,               # the edited file you import
  data = scientometrics,  # the original biblio data
  tag = 'ID'              # the field tag
)
}
