% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syn_export.R
\name{syn_export}
\alias{syn_export}
\title{Export a csv file to edit the synonyms table.}
\usage{
syn_export(data, tag = "", file = "")
}
\arguments{
\item{data}{a dataset}

\item{tag}{}

\item{file}{a file name as the export file. (notice: add th subtitle .csv)}
}
\value{
a csv file
}
\description{
Export a csv file to edit the synonyms table.
}
\examples{
syn_export(scientometrics, tag='ID', 'test1127.csv')

}
