% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syn_replace.R
\name{syn_replace}
\alias{syn_replace}
\title{replace synonyms by importing csv file}
\usage{
syn_replace(df, data, tag)
}
\arguments{
\item{df}{a dataframe includes synonyms and replaced terms}

\item{data}{a dataset}

\item{tag}{a field-tag, i.g. 'ID' or 'DE'.}
}
\value{
a dataframe
}
\description{
replace synonyms by importing csv file
}
\details{

}
\examples{
data(scientometrics, package = "bibliometrixData")
scientometrics$ID <- syn_replace( df = DTF,
data = scientometrics,
tag = 'ID')
}
